requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
pyTelegramBotAPI==4.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot-calendar==1.0.5
requests==2.28.1
SimpleQIWI==0.7
six==1.16.0
urllib3==1.26.12
uuid==1.30
flask